+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 26
     defaultConfig {
-        applicationId "com.ing.software.appscontrini2"
+        applicationId "com.ing.software.appscontrini"
         minSdkVersion 21
         targetSdkVersion 26
         versionCode 1
