+++ app/src/androidTest/java/com/ing/software/appscontrini/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.ing.software.appscontrini2;
+package com.ing.software.appscontrini;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
