+++ app/src/main/java/com/ing/software/appscontrini/OCR/ProbGrid.java
@@ -1,4 +1,4 @@
-package com.ing.software.appscontrini2.OCR;
+package com.ing.software.appscontrini.OCR;
 
 
 import java.util.HashMap;
