+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.ing.software.appscontrini2.MainActivity">
+    tools:context="com.ing.software.appscontrini.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
