+++ gradle_test.bat
@@ -1,0 +1,3 @@
+@echo off
+cd /d  "%~dp0"
+cmd /k gradle build
